tools/geneos: further clean-up of `ps`, revert PID/PPID to int and not int64
pkg/config: further clean-up to support changes above

+++ pkg/process/files.go
@@ -3,7 +3,7 @@ package process
 import "io/fs"
 
 type ProcessFDs struct {
-	PID   int64
+	PID   int
 	FD    int
 	Path  string
 	Lstat fs.FileInfo
